Add palette component
The component renders all the color combinations in the default theme.

As Tailwind classes are computed dynamically, we cannot use it in the usual way (all the non-static classes are being pruned).

It turns out that there is a simple way of integrating Tailwind to run on runtime instead, using the seemingly undocumented package `@tailwindcss/browser`. Which is nice, because using the compiler from `tailwindcss` directly is undocumented as well...

We do get a warning that you aren't supposed to use this in production, which can be ignored (but apparently not removed).